fix(queue): simplify nitro quickstart example
