--- packages/queue/examples/nitro/server/api/queues/welcome-payload.get.ts
@@ -1,11 +1,0 @@
-import type { WelcomeEmailPayload } from "../../queues/welcome-email"
-
-const samplePayload: WelcomeEmailPayload = {
-  email: "ava@example.com",
-  template: "default",
-}
-
-export default defineEventHandler(() => ({
-  queue: "welcome-email",
-  samplePayload,
-}))
