+++ packages/queue/examples/showcase.json
@@ -53,11 +53,9 @@
           "phases": {
             "configure": "nitro.config.ts",
             "define": "server/queues/welcome-email.ts",
-            "run": "server/api/queues/welcome.post.ts"
+            "run": "server/api/welcome.post.ts"
           },
           "supplementalFiles": [
-            "server/api/queues/welcome-defer.post.ts",
-            "server/api/queues/welcome-payload.get.ts",
             "package.json"
           ]
         },
@@ -65,11 +63,9 @@
           "phases": {
             "configure": "nitro.config.ts",
             "define": "server/queues/welcome-email.ts",
-            "run": "server/api/queues/welcome.post.ts"
+            "run": "server/api/welcome.post.ts"
           },
           "supplementalFiles": [
-            "server/api/queues/welcome-defer.post.ts",
-            "server/api/queues/welcome-payload.get.ts",
             "package.json"
           ]
         }
